Refactor into standalone middleware directory
