+++ controllers/user/index.js
@@ -3,7 +3,7 @@ const express = require('express');
 const router = express.Router();
 
 const auth = require('../../modules/authed');
-const middleware = require('../../modules/middleware-helpers');
+const cached = require('../../middleware/cached');
 const register = require('./register');
 const login = require('./login');
 const password = require('./password');
@@ -11,23 +11,23 @@ const dashboard = require('./dashboard');
 const { userBasePath, userEndpoints, emailPasswordValidations, formValidations } = require('./utils');
 
 // serve a logged-in user's dashboard
-router.get(userEndpoints.dashboard, auth.requireAuthed, middleware.noCache, dashboard.dashboard);
+router.get(userEndpoints.dashboard, auth.requireAuthed, cached.noCache, dashboard.dashboard);
 
 // register users
 router
     .route(userEndpoints.register)
     .get(auth.requireUnauthed, (req, res) => res.send('Temporarily removed.'))
-    // .get(auth.requireUnauthed, middleware.csrfProtection, register.registrationForm)
-    .post(emailPasswordValidations, middleware.csrfProtection, register.createUser);
+    // .get(auth.requireUnauthed, cached.csrfProtection, register.registrationForm)
+    .post(emailPasswordValidations, cached.csrfProtection, register.createUser);
 
 // login users
 router
     .route(userEndpoints.login)
-    .get(auth.requireUnauthed, middleware.csrfProtection, login.loginForm)
-    .post(middleware.csrfProtection, login.attemptAuth);
+    .get(auth.requireUnauthed, cached.csrfProtection, login.loginForm)
+    .post(cached.csrfProtection, login.attemptAuth);
 
 // logout users
-router.get(userEndpoints.logout, middleware.noCache, (req, res) => {
+router.get(userEndpoints.logout, cached.noCache, (req, res) => {
     req.logout();
     req.flash('justLoggedOut', true);
     req.session.save(() => {
@@ -36,18 +36,18 @@ router.get(userEndpoints.logout, middleware.noCache, (req, res) => {
 });
 
 // activate an account
-router.get(userEndpoints.activate, auth.requireAuthed, middleware.noCache, register.activateUser);
+router.get(userEndpoints.activate, auth.requireAuthed, cached.noCache, register.activateUser);
 
 // request a password reset email
 router
     .route(userEndpoints.requestpasswordreset)
-    .get(auth.requireUnauthed, middleware.noCache, password.requestResetForm)
+    .get(auth.requireUnauthed, cached.noCache, password.requestResetForm)
     .post(auth.requireUnauthed, formValidations.emailAddress, password.sendResetEmail);
 
 // change a password (with a token)
 router
     .route(userEndpoints.resetpassword)
-    .get(auth.requireUnauthed, middleware.noCache, password.changePasswordForm)
+    .get(auth.requireUnauthed, cached.noCache, password.changePasswordForm)
     .post(auth.requireUnauthed, formValidations.password, password.updatePassword);
 
 module.exports = router;
