+++ modules/boilerplate/static.js
@@ -1,4 +1,4 @@
-"use strict";
+'use strict';
 const app = require('../../server');
 const assets = require('../assets');
 const express = require('express');
@@ -6,6 +6,9 @@ const path = require('path');
 const config = require('config');
 
 // configure static files
-app.use('/' + assets.assetVirtualDir, express.static(path.join(__dirname, '../../public'), {
-    maxAge: config.get('staticExpiration')
-}));
+app.use(
+    '/' + assets.assetVirtualDir,
+    express.static(path.join(__dirname, '../../public'), {
+        maxAge: config.get('staticExpiration')
+    })
+);
